JUst simulations

+++ simulation_new/input_physical_data.h
@@ -1,5 +1,5 @@
 // Physical data of the system ( masses and distances of the 3 bodies )
-
-int initial_velocities();
-int initial_masses();
+double initial_velocities();
+double initial_masses();
 double initial_pos();
+
